Snowmobile in Cambridge Bay, Nunavut
ONC's icListen hydrophone in Cambridge Bay, NV picked up the sound of a snowmobile being driven across the frozen sea ice. You can distinctly hear the approach of the vehicle as well as shifts in the sea ice. Monitoring sea ice thickness is crucial for coastal communities in the Canadian Arctic so they know when it's safe to travel across the surface.